.requirements.txt
haystack-ai
tika